chore: update Release npm script with new project structure - also convert all requireJS to ESM (.mjs) files so that we can use newer import syntax and use latest deps
