+++ package.json
@@ -56,12 +56,9 @@
     "build:test": "node ./scripts/builds.mjs --test",
     "prebuild:prod": "npm run lint",
     "build:prod": "node ./scripts/builds.mjs --prod",
-    "prebuild:types": "npm run lint",
-    "build:types": "tsc --emitDeclarationOnly --incremental --declarationMap false --outDir dist/types",
-    "prebuild:types:prod": "npm run lint",
-    "build:types:prod": "tsc --emitDeclarationOnly --incremental --declarationMap --outDir dist/types",
-    "preview:release": "node ./scripts/release.js --create-release github --dry-run --skip-checks",
-    "release": "node ./scripts/release.js --create-release github",
+    "build:types": "tsc --emitDeclarationOnly --incremental --declarationMap --outDir dist/types",
+    "preview:release": "node ./scripts/release.mjs --create-release github --dry-run --skip-checks",
+    "release": "node ./scripts/release.mjs --create-release github",
     "presass:build": "rimraf dist/styles/css",
     "sass:build": "sass src/styles:dist/styles/css --style=compressed --quiet-deps --no-source-map",
     "postsass:build": "postcss dist/styles/css/**/* --dir dist/styles/css --base dist/styles/css --no-map --use cssnano --use autoprefixer --style=compressed",
@@ -74,8 +71,8 @@
   "devDependencies": {
     "@4tw/cypress-drag-drop": "^2.2.4",
     "@octokit/rest": "^19.0.13",
-    "@typescript-eslint/eslint-plugin": "^6.4.0",
-    "@typescript-eslint/parser": "^6.4.0",
+    "@typescript-eslint/eslint-plugin": "^6.6.0",
+    "@typescript-eslint/parser": "^6.6.0",
     "autoprefixer": "^10.4.15",
     "browser-sync": "^2.29.3",
     "chokidar": "^3.5.3",
@@ -86,27 +83,27 @@
     "cypress": "^12.17.4",
     "dotenv": "^16.3.1",
     "esbuild": "^0.19.2",
-    "eslint": "^8.47.0",
+    "eslint": "^8.48.0",
     "eslint-plugin-cypress": "^2.14.0",
-    "eslint-plugin-import": "^2.28.0",
+    "eslint-plugin-import": "^2.28.1",
     "eslint-plugin-node": "^11.1.0",
     "fs-extra": "^11.1.1",
     "git-url-parse": "^13.1.0",
-    "glob": "^10.3.3",
+    "glob": "^10.3.4",
     "http-server": "^14.1.1",
     "inquirer": "^9.2.10",
-    "new-github-release-url": "^1.0.0",
+    "load-json-file": "^7.0.1",
+    "new-github-release-url": "^2.0.0",
     "path": "^0.12.7",
     "picocolors": "^1.0.0",
-    "postcss": "^8.4.28",
+    "postcss": "^8.4.29",
     "postcss-cli": "^10.1.0",
     "rimraf": "^5.0.1",
-    "sass": "^1.65.1",
+    "sass": "^1.66.1",
     "semver": "^7.5.4",
     "servor": "^4.0.2",
     "strong-log-transformer": "^2.1.0",
-    "typescript": "^5.1.6",
-    "uglify-js": "^3.17.4",
+    "typescript": "5.1.6",
     "yargs": "^17.7.2"
   },
   "browserslist": [
