+++ tsconfig.json
@@ -31,6 +31,6 @@
       "./node_modules/@types"
     ]
   },
-  "exclude": ["node_modules", "**/*.spec.ts"],
+  "exclude": ["cypress", "node_modules", "**/*.spec.ts"],
   "include": ["**/*.ts", "index.d.ts", "**/*.d.ts"]
 }
